Я НЕ ПОНИМАТЬ
ДЕлал Димка

+++ Assets/Scripts/Block.cs
@@ -5,7 +5,7 @@ using UnityEngine;
 [System.Serializable]
 public class Block {
     public string display_name;
-    public byte id;
+    public int id;
     public Sprite[] sprites;
     public bool isSolid = true;
     public Drop[] drops;
